refactorizacion del map
