+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@gorhom/bottom-sheet": "^4.4.5",
     "@notifee/react-native": "^7.0.1",
-    "@react-native-community/geolocation": "^3.0.4",
+    "@react-native-community/geolocation": "^3.0.5",
     "@react-native-community/netinfo": "^9.3.7",
     "@react-native-firebase/app": "^16.4.6",
     "@react-native-firebase/auth": "^16.4.6",
@@ -38,7 +38,7 @@
     "react-native-gesture-handler": "^2.8.0",
     "react-native-image-picker": "^4.10.0",
     "react-native-maps": "^1.3.2",
-    "react-native-pager-view": "^6.1.1",
+    "react-native-pager-view": "^6.1.4",
     "react-native-paper": "^4.12.5",
     "react-native-permissions": "^3.6.1",
     "react-native-reanimated": "^2.13.0",
@@ -48,7 +48,7 @@
     "react-native-splash-screen": "^3.3.0",
     "react-native-tab-view": "^3.3.0",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-vision-camera": "^2.15.2",
+    "react-native-vision-camera": "^2.15.4",
     "react-redux": "^8.0.5",
     "yup": "^0.32.11"
   },
