+++ src/Store/DataFirebase/FirebaseSlice.tsx
@@ -1,0 +1,46 @@
+
+// import { createSlice } from '@reduxjs/toolkit';
+// import { PermissionStatus, request,PERMISSIONS } from 'react-native-permissions';
+// import { Platform } from 'react-native';
+
+
+
+
+// export interface  TypeData{
+//     contraseña:number,
+//     correo:string,
+//     nombre:string,
+//     photoPerfil:string,
+//     hora:string,
+//     latitude:string,
+//     longitude:string,
+//     nombreUser:string,
+//     photo:string
+// }
+
+// export const initialState :TypeData = {
+//  contraseña:0,
+//  correo:"",
+//  nombre:"",
+//  photoPerfil:"",
+//  hora:"",
+//  latitude:"",
+//  longitude:"",
+//  nombreUser:"",
+//  photo:""
+// }
+ 
+
+
+
+
+// export const FirebaseSlice = createSlice({
+//     name: "Firebase",
+//     initialState,
+//     reducers: {
+//          obtenerAreglo:(state,payload)=>{
+
+//          }}
+// })
+
+// export const { GiveCoreoForm } = FirebaseSlice.actions;
